[MODFEE-83] Post-merge adjustments
