+++ src/test/java/org/folio/test/support/EntityBuilder.java
@@ -27,9 +27,4 @@ public class EntityBuilder {
       .withStatus(new Status().withName("Open"));
   }
 
-  public static Account createAccount(double amount) {
-    return createAccount()
-      .withAmount(amount)
-      .withRemaining(amount);
-  }
 }
